Extract string and constant handling from expression.rs
